+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 ### Back-end
 - Added current equipment tracking
 - Added automatic backups for config and data.
+- Removed Client-side Eman 9 validation